Add pubchem and macs fingerprint data
I altered the variational autoencoder to get it too compile with a higher efficiency and to work with a later version of Python (3.7). I also added the pubchem fingerprint data from the fingerprint sort algorithm for the CCLE dataset. I also altered the training model to work with the CCLE dataset (x_train) instead of the TCGA gene expression dataset (X_train).